# openapi.generated2.yaml
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
paths:
...
  /api/v1/response-sets/{response_set_id}/screens/{screen_key}:
    get:
      tags:
      - ScreenView
      - Gating
      - ScreenView
      summary: Get a screen with its questions and any existing answers
      operationId: getScreenWithAnswers
      parameters:
      - name: response_set_id
        in: path
        required: true
        schema:
          type: string
          title: Response Set Id
      - name: screen_key
        in: path
        required: true
        schema:
          type: string
          title: Screen Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScreenViewEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/response-sets/{id}/regenerate-check:
    post:
      tags:
      - ScreenView
      - Gating
      - Gating
      summary: Check whether generation may proceed (gating)
      operationId: regenerateCheck
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/authoring/questionnaires/{questionnaire_id}/screens/{screen_id}:
    patch:
      tags:
      - Authoring
      summary: Update Screen
      operationId: update_screen_api_v1_authoring_questionnaires__questionnaire_id__screens__screen_id__patch
      parameters:
      - name: questionnaire_id
        in: path
        required: true
        schema:
          type: string
          title: Questionnaire Id
      - name: screen_id
        in: path
        required: true
        schema:
          type: string
          title: Screen Id
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ScreenAlias:
      properties:
        screen_key:
          type: string
          title: Screen Key
        screen_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Screen Id
      type: object
      required:
      - screen_key
      title: ScreenAlias
    ScreenView:
      properties:
        screen_key:
          type: string
          title: Screen Key
        questions:
          items: {}
          type: array
          title: Questions
        etag:
          anyOf:
          - type: string
          - type: 'null'
          title: Etag
      type: object
      required:
      - screen_key
      - questions
      title: ScreenView
    ScreenViewEnvelope:
      properties:
        screen_view:
          $ref: '#/components/schemas/ScreenView'
        screen:
          anyOf:
          - $ref: '#/components/schemas/ScreenAlias'
          - type: 'null'
      type: object
      required:
      - screen_view
      title: ScreenViewEnvelope
      description: 'Envelope for GET screen responses.


        Matches the contract expected by integration: include both

        `screen_view` and a `screen` alias object.'
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError